words_tag_cloud/requirements.txt
aiohttp==3.6.3
async-timeout==3.0.1
asyncio==3.4.3
attrs==20.2.0
certifi==2020.6.20
chardet==3.0.4
click==7.1.2
cycler==0.10.0
idna==2.10
joblib==0.17.0
kiwisolver==1.2.0
lxml==4.6.1
matplotlib==3.3.2
multidict==4.7.6
nltk==3.5
numpy==1.19.2
Pillow==8.0.0
pyparsing==2.4.7
python-dateutil==2.8.1
regex==2020.10.15
requests==2.24.0
six==1.15.0
tqdm==4.50.2
typing-extensions==3.7.4.3
urllib3==1.25.10
wordcloud==1.8.0
yarl==1.5.1
